Seperation of request and access token + refactoring
The storages themselves don't trigger if the access token is missing.
This enables a TokenNotFoundException if you try to do a request with
only a request token.
